Make changes to satisfy dialyzer and add some tests
This commit eliminates the problems detected by `dialyzer`.

I added a doctest to illustrate the "happy path," and added a @spec
annotation so that the input and output types appear in documentation.

The test suite was updated with a couple of "useful" tests.

One thing I noticed was that Timex.parse!/1 happily parsed "2015-02-29",
but Erlang's :calendar.valid_date({2015, 2, 29}) returns false. That's
not in the scope of this exercise to fix!